refactor(Output): extract to Core
